requirements.txt
pynput
web3
ruamel.yaml
selenium
